RAT-362: Refactorings during review and changelog
