+++ apache-rat-core/src/test/java/org/apache/rat/ReporterOptionsTest.java
@@ -24,11 +24,9 @@ import java.nio.file.Path;
 import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
-import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.rat.api.RatException;
 import org.apache.rat.report.claim.ClaimStatistic;
 import org.apache.rat.test.AbstractConfigurationOptionsProvider;
-import org.apache.rat.test.AbstractOptionsProvider;
 import org.apache.rat.testhelpers.FileUtils;
 import org.apache.rat.testhelpers.XmlUtils;
 import org.apache.rat.utils.DefaultLog;
